Map Demo
This is an increment in login.
Map module is added.
when we login we see a button on success page. "Show Map"
when we press the button a controller is called which fetch data from locations table.
and passes this data to view which displays map and mark some points.  we get latitude and langtiude from model.