Improvements in dashboard and multi select
Changed dashboard for the new format of data. Created functions for the the multiselect data. Made changes in the Householdmodel and Householdcontroller file.
For the multiselect we are getting data from each column and then comibing them in the array of objects.
In database structure of table household is changed. its attribute names are changed.
Queries for each question is updated with respect to new column name.